feat(admin): 後台導航重構 — 側邊欄分組 + 儀表板精簡 + 球種來源拆分
- AdminLayoutClient 從頂部 8 項扁平導航改為左側側邊欄，分 3 大區塊（總覽/內容/設定）
- 手機端側邊欄收合，漢堡選單 overlay 展開
- 儀表板移除寫手陣容、發布頻道、自動化面板，只保留統計卡片和爬蟲狀態
- /admin/sports 拆分為球種分類 + /admin/sources 爬蟲來源管理
- 新增 /admin/pipeline 和 /admin/review placeholder 頁面
