+++ src/app/admin/pipeline/page.tsx
@@ -1,0 +1,13 @@
+export default function PipelinePage() {
+  return (
+    <div className="space-y-4">
+      <h1 className="text-2xl font-bold">管線狀態</h1>
+      <div className="flex items-center justify-center py-24 text-gray-400 border-2 border-dashed rounded-lg">
+        <div className="text-center">
+          <p className="text-lg font-medium">Coming Soon</p>
+          <p className="text-sm mt-1">視覺化管線流程圖即將推出（Issue #163）</p>
+        </div>
+      </div>
+    </div>
+  );
+}
