+++ src/app/admin/review/page.tsx
@@ -1,0 +1,13 @@
+export default function ReviewPage() {
+  return (
+    <div className="space-y-4">
+      <h1 className="text-2xl font-bold">審稿佇列</h1>
+      <div className="flex items-center justify-center py-24 text-gray-400 border-2 border-dashed rounded-lg">
+        <div className="text-center">
+          <p className="text-lg font-medium">Coming Soon</p>
+          <p className="text-sm mt-1">手動覆審工作台即將推出（Issue #164）</p>
+        </div>
+      </div>
+    </div>
+  );
+}
